IGNITE-26371 Refactor base abstract test (better support tests with extensions)
